Set components new filename
