+++ src/components/atoms/SliderComp.jsx
