+++ src/components/layouts/admin/LayoutComp.jsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
 import { useAuth } from '@hooks/use-auth';
-import Header from '../Header';
-import Main from './Main';
+import Header from '../HeaderComp';
+import Main from './MainComp';
 
 export default function AdminLayout() {
   const navigate = useNavigate();
