+++ src/components/layouts/admin/MainComp.jsx
@@ -1,4 +1,4 @@
-import { Sidebar } from './Sidebar';
+import { Sidebar } from './SidebarComp';
 
 export default function Main(props) {
   const { children } = props;
