+++ src/pages/SearchPage.jsx
@@ -3,8 +3,8 @@ import { useSearchParams } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { Button } from 'antd';
 import { SearchOutlined } from '@ant-design/icons';
-import { Select } from '@atoms/Select';
-import { Slider } from '@atoms/Slider';
+import { Select } from '@atoms/SelectComp';
+import { Slider } from '@atoms/SliderComp';
 import { EmptyState } from '@atoms/EmptyState';
 import { ProductCard } from '@organisms/ProductCard';
 import { getProductBySearch } from '@services/products';
